+++ src/restart.h
@@ -15,7 +15,7 @@ extern "C" {
 #endif
 
 struct timeval add2currenttime(double seconds);
-int copyfile(int fromfd, int tofd);
+size_t copyfile(int fromfd, int tofd);
 int r_close(int fildes);
 int r_dup2(int fildes, int fildes2);
 int r_open2(const char *path, int oflag);
